backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15
python-dotenv==1.0.0
